+++ priv/repo/migrations/20210924113308_add_starting_balance_to_book.exs
@@ -3,7 +3,7 @@ defmodule BudgetTrackingTool.Repo.Migrations.AddStartingBalanceToBook do
 
   def change do
     alter table(:books) do
-      add :starting_balance, :float, default: 0
+      add :starting_balance, :integer, default: 0
     end
   end
 end
